LinStatExpCov {libcoin}	R Documentation

Linear Statistics with Expectation and Covariance

Description

Strasser-Weber type linear statistics and their expectation and covariance under the independence hypothesis

Usage

LinStatExpCov(X, Y, ix = NULL, iy = NULL, weights = integer(0),
              subset = integer(0), block = integer(0), checkNAs = TRUE,
              varonly = FALSE, nresample = 0, standardise = FALSE,
              tol = sqrt(.Machine$double.eps))
lmult(x, object)

Arguments

X
numeric matrix of transformations.

Y
numeric matrix of influence functions.

ix
an optional integer vector expanding X.

iy
an optional integer vector expanding Y.

weights
an optional integer vector of non-negative case weights.

subset
an optional integer vector defining a subset of observations.

block
an optional factor defining independent blocks of observations.

checkNAs
a logical for switching off missing value checks. This included switching off checks for suitable values of subset. Use at your own risk.

varonly
a logical asking for variances only.

nresample
an integer defining the number of permuted statistics to draw.

standardise
a logical asking to standardise the permuted statistics.

tol
tolerance for zero variances.

x
a contrast matrix to be left-multiplied in case X was a factor.

object
an object of class "LinStatExpCov".

Details

This function implements the permutation test framework by Strasser and Weber (1999), see also Hothorn, Hornik, van de Wiel, and Zeileis (2006) and Hothorn, Hornik, van de Wiel, and Zeileis (2008).

The function, after minimal preprocessing, calls the underlying C code and computes the linear statistic, its expectation and covariance and, optionally, nresample samples from its permutation distribution.

When both ix and iy are missing, the number of rows of X and Y is the same, ie the number of observations.

When X is missing and ix a factor, the code proceeds as if X were a dummy matrix of ix without explicitly computing this matrix.

Both ix and iy being present means the code treats them as subsetting vectors for X and Y. Note that ix = 0 or iy = 0 means that the corresponding observation is missing and the first row or X and Y must be zero.

lmult allows left-multiplication of a contrast matrix when X was (equivalent to) a factor.

Value

A list.

References

Hothorn T., Hornik K., van de Wiel M. A., Zeileis A. (2006). “A Lego System for Conditional Inference.” The American Statistician, 60(3), 257–263. doi: 10.1198/000313006X118430.

Hothorn T., Hornik K., van de Wiel M. A., Zeileis A. (2008). “Implementing a Class of Permutation Tests: The coin Package.” Journal of Statistical Software, 28(8), 1–23. doi: 10.18637/jss.v028.i08.

Strasser H., Weber C. (1999). “On the Asymptotic Theory of Permutation Statistics.” Mathematical Methods of Statistics, 8, 220-250. Preprint available from https://epub.wu.ac.at/102.

Examples

wilcox.test(Ozone ~ Month, data = airquality, subset = Month %in% c(5, 8),
            exact = FALSE, correct = FALSE)

aq <- subset(airquality, Month %in% c(5, 8))
X <- as.double(aq$Month == 5)
Y <- as.double(rank(aq$Ozone, na.last = "keep"))
doTest(LinStatExpCov(X, Y))

[Package libcoin version 1.0-12 Index]